Updated sdk structure and a new module for extraction of near addresses from strings
