+++ apps/buildbox/widget/page/view.jsx
@@ -4,6 +4,12 @@ const { User, Button } = VM.require("buildhub.near/widget/components") || {
   Button: () => <></>,
 };
 
+const { extractValidNearAddresses } = VM.require(
+  "buildbox.near/widget/utils.projects-sdk"
+) || {
+  extractValidNearAddresses: () => {},
+};
+
 const data = Social.get(id + "/**", "final");
 
 if (!id || !data) {
@@ -179,46 +185,48 @@ const {
   learning,
 } = JSON.parse(data[""]);
 
-/* HELPER FUNCTION */
-function isNearAddress(address) {
-  if (typeof address !== "string") {
-    return false;
-  }
+// /* HELPER FUNCTION */
+// function isNearAddress(address) {
+//   if (typeof address !== "string") {
+//     return false;
+//   }
 
-  // Allow both ".near" and ".testnet" endings
-  if (!address.endsWith(".near") && !address.endsWith(".testnet")) {
-    return false;
-  }
+//   // Allow both ".near" and ".testnet" endings
+//   if (!address.endsWith(".near") && !address.endsWith(".testnet")) {
+//     return false;
+//   }
 
-  const parts = address.split(".");
-  if (parts.length !== 2) {
-    return false;
-  }
+//   const parts = address.split(".");
+//   if (parts.length !== 2) {
+//     return false;
+//   }
 
-  if (parts[0].length < 2 || parts[0].length > 32) {
-    return false;
-  }
+//   if (parts[0].length < 2 || parts[0].length > 32) {
+//     return false;
+//   }
 
-  if (!/^[a-z0-9_-]+$/i.test(parts[0])) {
-    return false;
-  }
+//   if (!/^[a-z0-9_-]+$/i.test(parts[0])) {
+//     return false;
+//   }
 
-  return true;
-}
+//   return true;
+// }
 
 function Team({ members }) {
   if (members) {
     // removing extra characters and splitting the string into an array
-    const arr = members.replace(/[\[\]\(\)@]/g, "").split(/[\s,]+/);
+    // const arr = members.replace(/[\[\]\(\)@]/g, "").split(/[\s,]+/);
 
-    // filtering out teammates that are not near addresses
-    const hexRegex = /^[0-9A-F\-_]+$/i;
-    const valid = arr.filter((teammate) => {
-      if (hexRegex.test(teammate)) {
-        return teammate;
-      }
-      return isNearAddress(teammate);
-    });
+    // // filtering out teammates that are not near addresses
+    // const hexRegex = /^[0-9A-F\-_]+$/i;
+    // const valid = arr.filter((teammate) => {
+    //   if (hexRegex.test(teammate)) {
+    //     return teammate;
+    //   }
+    //   return isNearAddress(teammate);
+    // });
+
+    const valid = extractValidNearAddresses(members);
 
     const extractNearAddress = (id) => {
       const parts = id.split("/");
